\docType{package}
\name{cdtRmodelling-package}
\alias{cdtRmodelling}
\alias{cdtRmodelling-package}
\title{R courses for Newcastle University's CDT in Cloud Computing for Big Data}
\description{
  Functions and datasets used for Newcastle University's 
  CDT in Cloud Computing for Big Data. This package is 
  used in the "Statistical Modelling" course.
}
\keyword{package}

